[Models] Move models with migrations to appropriate apps
